pyproject.toml
[project]
name = "colonia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    # Using a specific tag
    "rupy-api >=v0.1.7",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.27",
    "alembic>=1.17.2",
    "pika>=1.3.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "pytest>=9.0.2",
]
